Reorganize + add settings.py
